Agent/requirements.txt
# Multi-Agent Collaborative System - Complete Dependencies
# Compatible with LLM environments (Ollama, Qwen, etc.)

# Core LLM and AI Dependencies
openai>=1.3.0                 # OpenAI-compatible API client for Qwen/Ollama
pydantic>=2.0.0               # Data validation and parsing using BaseModel
python-dotenv>=1.0.0          # Environment variable management

# Web and HTTP Dependencies  
requests>=2.31.0              # HTTP requests for APIs and web search
httpx>=0.24.0                 # Modern async HTTP client
beautifulsoup4>=4.12.0        # HTML parsing for web scraping
lxml>=4.9.0                   # XML/HTML parser backend

# Web Framework (for Flask demo app)
flask>=2.3.0                  # Web framework for arXiv daily demo
jinja2>=3.1.0                 # Template engine for Flask
markupsafe>=2.1.0             # Safe string handling for templates

# arXiv API Integration
arxiv>=1.4.0                  # Official arXiv API client
feedparser>=6.0.0             # RSS/Atom feed parsing

# Logging and Output
loguru>=0.7.0                 # Advanced logging with rich features
rich>=13.0.0                  # Rich terminal output and formatting

# File and Async Operations
aiofiles>=23.0.0              # Async file operations

# Development and Testing
pytest>=7.4.0                # Testing framework
pytest-asyncio>=0.21.0       # Async testing support
black>=23.0.0                 # Code formatting
flake8>=6.0.0                 # Code linting